fix(components): move to utils folder
